add OrderLineService, OrderLineController

+++ backend/src/main/java/pl/sdacademy/repositories/OrderLineRepository.java
@@ -3,5 +3,5 @@ package pl.sdacademy.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import pl.sdacademy.entities.OrderLine;
 
-public interface OrderLineRepository extends JpaRepository<OrderLine,Integer> {
+public interface OrderLineRepository extends JpaRepository<OrderLine,Long> {
 }
